+++ CHANGELOG.md
@@ -72,6 +72,7 @@ ______________________________________________________________________
 - `getReductionParameters()` also takes a `utc_date` argument and will re-calculate if the value in the KVS doesn't exist or doesn't match (see #49 and !80)
 - Attached `Measurement` to `Observation` object (see #156 and !141)
 - bump `mjolnir` version to 1.1.3
+- `MeasurementType.calculate()`, `Measurement.calculateMeasurement()`, and `getSlantRangeVector()` all take sensor/target ECI states and UTC as parameters (see #158)
 
 ### Deprecated